chroot: /etc/ssh/ might not be populated

+++ scripts/chroot.sh
@@ -846,9 +846,11 @@ echo "${rfs_username}:${rfs_password}" > /tmp/user_password.list
 sudo mv /tmp/user_password.list ${DIR}/deploy/${export_filename}/user_password.list
 
 #Fixes:
-#Remove pre-generated ssh keys, these will be regenerated on first bootup...
-sudo rm -rf ${tempdir}/etc/ssh/ssh_host_* || true
-sudo touch ${tempdir}/etc/ssh/ssh.regenerate || true
+if [ -d ${tempdir}/etc/ssh/ ] ; then
+	#Remove pre-generated ssh keys, these will be regenerated on first bootup...
+	sudo rm -rf ${tempdir}/etc/ssh/ssh_host_* || true
+	sudo touch ${tempdir}/etc/ssh/ssh.regenerate || true
+fi
 
 #extra home, from chroot machine when running npm install xyz:
 unset extra_home
